requirements.txt
django==5.0.2
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.27.1
python-dotenv==1.0.1
django-jazzmin==3.0.1
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1
